+++ inquisitor/src/engine/sdl/CSDL.cpp
@@ -13,7 +13,7 @@ CSDL::CSDL()
 	SDL_GetVersion( &version_linked );
 	SDL_VERSION( &version_compiled );
 
-	logDEBUG( "SDL has version '{0}.{1}.{2}'", static_cast< unsigned short >( version_linked.major ), static_cast< unsigned short >( version_linked.minor ), static_cast< unsigned short >( version_linked.patch ) );
+	logDEBUG( "SDL has version '{0}.{1}.{2}'", version_linked.major, version_linked.minor, version_linked.patch );
 
 	if( SDL_COMPILEDVERSION != SDL_VERSIONNUM( version_linked.major, version_linked.minor, version_linked.patch ) )
 	{
